+++ packages/@tinacms/cli/src/cmds/init/setup-files/index.ts
@@ -10,40 +10,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import chalk from 'chalk'
-export const adminPage = `import { TinaAdmin } from 'tinacms';
-export default TinaAdmin;
-`
-
-export const blogPost = `---
-title: Vote For Pedro
----
-# Welcome to the blog.
-
-> To edit this site head over to the [\`/admin\`](/admin) route. Then click the pencil icon in the bottom lefthand corner to start editing 🦙.
-
-# Dixi gaude Arethusa
-
-<PageSection heading="Oscula mihi" content="Lorem markdownum numerabilis armentorum platanus, cultros coniunx sibi per
-silvas, nostris clausit sequemur diverso scopulosque. Fecit tum alta sed non
-falcato murmura, geminas donata Amyntore, quoque Nox. Invitam inquit, modo
-nocte; ut ignis faciemque manes in imagine sinistra ut mucrone non ramos
-sepulcro supplex. Crescentesque populos motura, fit cumque. Verumque est; retro
-sibi tristia bracchia Aetola telae caruerunt et."/>
-
-
-## Mutato fefellimus sit demisit aut alterius sollicito
-
-Phaethonteos vestes quem involvite iuvenca; furiali anne: sati totumque,
-**corpora** cum rapacibus nunc! Nervis repetatne, miserabile doleas, deprensum
-hunc, fluctus Threicio, ad urbes, magicaeque, quid. Per credensque series adicis
-poteram [quidem](#)! Iam uni mensas victrix
-vittas ut flumina Satyri adulter; bellum iacet domitae repercusso truncis urnis
-mille rigidi sub taurum.
-
-
-`
-
 export const nextPostPage = ({
   usingSrc,
 }: {
@@ -303,47 +269,3 @@ export const nextPostPage = ({
       </div>
     )
   }`
-
-export const AppJsContent = (usingSrc: boolean, extraImports?: string) => {
-  const importLine = `import Tina from '${
-    usingSrc ? '../' : ''
-  }../.tina/components/TinaDynamicProvider.js'`
-
-  return `${importLine}
-${extraImports || ''}
-
-const App = ({ Component, pageProps }) => {
-  return (
-    <Tina>
-      <Component {...pageProps} />
-    </Tina>
-  )
-}
-
-export default App
-`
-}
-export const AppJsContentPrintout = (
-  usingSrc: boolean,
-  extraImports?: string
-) => {
-  const importLine = chalk.green(
-    `+ import Tina from '${
-      usingSrc ? '../' : ''
-    }../.tina/components/TinaDynamicProvider.js'`
-  )
-
-  return `${importLine}
-  ${extraImports || ''}
-
-  const App = ({ Component, pageProps }) => {
-  return (
-    ${chalk.green('+ <Tina>')}
-        <Component {...pageProps} />
-    ${chalk.green('+ </Tina>')}
-  )
-}
-
-export default App
-`
-}
